Prepared the code handling all incoming digest types. Added support for listeners.

+++ src/main/java/com/gossiperl/client/Util.java
@@ -1,7 +1,7 @@
 package com.gossiperl.client;
 
 public class Util {
-    public static int getTimestamp() {
-        return (int)System.currentTimeMillis()/1000;
+    public static long getTimestamp() {
+        return (long)System.currentTimeMillis()/1000;
     }
 }
